Improve pub-related names and packages
