+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/publish/PublishingExtensions.kt
